refactor(core): split database module tests
